finish uploads clips back-end

+++ src/userphoto/userphoto.service.ts
@@ -20,7 +20,7 @@ export class PhotoUserService {
         const photos = await this.photoRepo.find({ relations: {
             user: true,
             likesPhoto: true
-        }, where: { user: { id: userId } }})
+        }, where: { user: { id: userId } }, order: { id: 'DESC' }})
         
         return photos
     }
@@ -29,11 +29,9 @@ export class PhotoUserService {
         const photos = await this.photoRepo.find({ relations: {
             user: true,
             likesPhoto: true
-        }, where: { user: { id: userId } }})
+        }, where: { user: { id: userId } }, take: 3, order: { id: 'DESC' }})
 
-        const arr = photos.sort((a, b) => a.id - b.id).reverse()
-
-        return this.maxPhotos(arr, 3)
+        return photos
     }
 
     async getOnePhoto(id: number) {
@@ -48,20 +46,18 @@ export class PhotoUserService {
     }
 
     async deletePhotos(id: number) {
-        const photo = await this.photoRepo.findOneBy({ id })
+        const photo = await this.photoRepo.findOne({ relations: { 
+            commentsPhoto: true,
+            likesPhoto: true,
+            user: true 
+        }, where: { id } })
 
+        if(!photo) throw new HttpException('Фото не найдено', HttpStatus.NOT_FOUND)
+
+        await this.photoRepo.delete({ id })
         const file = path.join(__dirname, `../../uploads/userPhoto/${photo.photo}`)
         fs.unlink(file, err => console.log(err))
-        return this.photoRepo.delete(id)
+        return { delete: true }
     }
 
-    async maxPhotos(arr: any, count: number) {
-       let result = []
-
-       for(let i = 0; i < count; i++) {
-          if(arr[i]) result.push(arr[i]) 
-       }
-
-       return result
-    }
 }
